+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 * Add support for `update_secret` for oauth2 authentication module
 * Add support for TLS "identity" (client certificate)
+* Consumer can now be used as an `Iterator<Item = Delivery>`
 
 ### 0.26.11 (2019-09-17)